backend/requirements.txt
fastapi==0.129.1
uvicorn[standard]==0.41.0
python-multipart==0.0.22
PyMuPDF==1.27.1
chromadb==1.5.1
sentence-transformers==5.2.3
pydantic==2.12.5
rank-bm25==0.2.2

# Table Extraction pipeline
pdfplumber==0.11.0
camelot-py==0.11.0
pandas==2.2.1

# Database & Queues (Production Scalability)
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
greenlet==3.0.3
redis==5.0.3
celery==5.3.6

# LLM support (Enabled for advanced parsing fallback)
gpt4all==2.8.2
